refactor: FSD — model/api/ui для фич, компоненты для editor
features/create-snippet:
  - выделены api/createSnippet.ts, model/useCreateSnippet.ts
  - React-хуки исправлены (были на уровне модуля)
  - TSDoc + TODO про синхронизацию языков

features/manage-snippets:
  - api/deleteSnippet.ts, api/createExampleSnippet.ts
  - model/useManageSnippets.ts (выделение, удаление, пример)
  - model/useSnippetFilter.ts (загрузка, поиск, фильтр, сортировка)

features/run-code:
  - model/useRunner.ts (консоль, запуск, хоткей Ctrl+Enter)

features/share-snippet:
  - types.ts, copyToClipboard -> shared/lib/clipboard.ts

pages/editor:
  - вынесены EditorHeader, EditorSidebar, EditorStatusBar
  - useSession/useAuthModal перенесены в EditorHeader
  - 318 строк (было 607)

pages/dashboard:
  - вся логика заменена на useSnippetFilter + useManageSnippets
  - удалены неиспользуемые импорты

shared/lib:
  - добавлен clipboard.ts
