+++ frontend/src/v2/app/providers/index.ts
@@ -1,0 +1,2 @@
+export { SessionProvider } from './SessionProvider';
+export { AuthModalProvider } from './AuthModalProvider'
