+++ frontend/src/v2/entities/snippet/index.ts
@@ -1,3 +1,3 @@
 export type { Snippet } from './types';
-export { sampleCode } from './sampleCode';
-export { SNIPPETS_QUERY_KEY } from './api';
+export { sampleCode } from './lib/sampleCode';
+export { SNIPPETS_QUERY_KEY } from './lib/constants';
