+++ frontend/src/v2/entities/snippet/lib/constants.ts
