+++ frontend/src/v2/entities/snippet/lib/sampleCode.ts
