+++ frontend/src/v2/entities/user/model/session.ts
@@ -1,0 +1,13 @@
+import {
+  createContext,
+  useContext,
+} from 'react';
+import type { SessionContextValue } from '../types';
+
+export const SessionContext = createContext<SessionContextValue | null>(null);
+
+export function useSession(): SessionContextValue {
+  const ctx = useContext(SessionContext);
+  if (!ctx) throw new Error('useSession must be used within SessionProvider');
+  return ctx;
+}
