+++ frontend/src/v2/features/auth/index.ts
@@ -1,1 +1,1 @@
-export { useAuthModal, AuthModalProvider} from './ui/AuthModal';
+export { useAuthModal } from './model/authModal';
