+++ frontend/src/v2/features/auth/lib/constants.ts
@@ -1,4 +1,4 @@
-import { type AuthMode } from '../context';
+import { type AuthMode } from '../model/authModal';
 
 export const titles: Record<AuthMode, string> = {
   login: 'Вход в Runit',
