+++ frontend/src/v2/features/auth/model/authModal.ts
@@ -1,14 +1,6 @@
 import { createContext, useContext } from 'react';
+import { AuthModalContextValue } from '../types'
 
-export type AuthMode = 'login' | 'register' | 'reset';
-
-export type AuthModalContextValue = {
-  open: (mode?: AuthMode) => void;
-  close: () => void;
-  mode: AuthMode;
-  setMode: (mode: AuthMode) => void;
-  opened: boolean;
-};
 
 export const AuthModalContext = createContext<AuthModalContextValue | null>(null);
