+++ frontend/src/v2/features/auth/ui/ResetForm.tsx
@@ -8,7 +8,7 @@ import {
 } from '@mantine/core';
 import { useForm } from '@mantine/form';
 import { notifications } from '@mantine/notifications';
-import { useAuthModal } from '../context';
+import { useAuthModal } from '../model/authModal';
 import FormHeader from './FormHeader';
 import { validateEmail } from '../lib/validate';
 import { titles } from '../lib/constants';
