+++ frontend/src/v2/features/create-snippet/lib/constants.ts
@@ -1,0 +1,6 @@
+/** Подсказки для каждого уровня видимости сниппета. */
+export const VISIBILITY_HINTS: Record<string, string> = {
+  private: 'Виден только вам',
+  link: 'Доступен всем, у кого есть ссылка',
+  public: 'Виден в вашем профиле и в поиске',
+};
