+++ frontend/src/v2/features/manage-snippets/api/deleteSnippet.ts
@@ -1,0 +1,5 @@
+import type { TrpcClient } from '../../../shared/api/trpc'
+
+/** Удаляет сниппет по id. */
+export const deleteSnippet = (trpc: TrpcClient, id: number) =>
+  trpc.snippets.deleteSnippet.mutate({ id })
