+++ frontend/src/v2/features/manage-snippets/index.ts
@@ -1,3 +1,5 @@
 export { default as SnippetCard } from './ui/SnippetCard';
 export { default as EmptyState } from './ui/EmptyState';
 export { default as BulkBar } from './ui/BulkBar';
+export { default as useManageSnippets } from './model/useManageSnippets';
+export { default as useSnippetFilter } from './model/useSnippetFilter'
