+++ frontend/src/v2/features/manage-snippets/lib/constants.ts
@@ -1,0 +1,2 @@
+/** Языки, доступные для создания сниппета-примера в пустом состоянии дашборда. */
+export const EXAMPLE_LANGS = ['javascript', 'python', 'php'] as const;
