+++ frontend/src/v2/features/run-code/index.ts
@@ -1,2 +1,3 @@
 export { default as ConsolePanel } from './ui/ConsolePanel';
-export type { OutputTab } from './ui/ConsolePanel';
+export type { OutputTab } from './types';
+export { default as useRunner } from './model/useRunner'
