+++ frontend/src/v2/features/share-snippet/types.ts
@@ -1,0 +1,8 @@
+/** Свойства модального окна «Поделиться сниппетом». */
+export type Props = {
+  opened: boolean;
+  onClose: () => void;
+  username: string;
+  slug: string;
+  saved: boolean;
+};
