+++ frontend/src/v2/shared/lib/index.ts
@@ -1,2 +1,3 @@
 export { plural, relativeDate } from './dates';
 export { initialsOf } from './initialsOf';
+export { copyToClipboard } from './clipboard';
